Fixed numerous MSVC warnings (#76)

+++ benchmark/extent_benchmark_weird.cc
@@ -48,7 +48,7 @@ class IndexBenchmark {
     void QueryWorld(benchmark::State& state);
 
     const TestGenerator data_type_;
-    const int num_entities_;
+    const size_t num_entities_;
 
     PhTree<DIM, int> tree_;
     std::default_random_engine random_engine_;
@@ -81,8 +81,8 @@ template <dimension_t DIM>
 void IndexBenchmark<DIM>::SetupWorld(benchmark::State& state) {
     logging::info("Setting up world with {} entities and {} dimensions.", num_entities_, DIM);
     CreatePointData<DIM>(points_, data_type_, num_entities_, 0, GLOBAL_MAX);
-    for (int i = 0; i < num_entities_; ++i) {
-        tree_.emplace(points_[i], i);
+    for (size_t i = 0; i < num_entities_; ++i) {
+        tree_.emplace(points_[i], (int)i);
     }
 
     state.counters["total_result_count"] = benchmark::Counter(0);
