+++ benchmark/query_mm_d_benchmark.cc
@@ -161,7 +161,7 @@ struct CounterMultiMap {
 };
 
 template <dimension_t DIM, Scenario SCENARIO>
-typename std::enable_if<SCENARIO == Scenario::TREE_WITH_MAP, int>::type CountEntries(
+typename std::enable_if<SCENARIO == Scenario::TREE_WITH_MAP, size_t>::type CountEntries(
     TestMap<Scenario::TREE_WITH_MAP, DIM>& tree, const Query& query) {
     CounterTreeWithMap counter{query.center, query.radius, 0};
     tree.for_each(query.box, counter);
@@ -169,14 +169,14 @@ typename std::enable_if<SCENARIO == Scenario::TREE_WITH_MAP, int>::type CountEnt
 }
 
 template <dimension_t DIM, Scenario SCENARIO>
-int CountEntries(TestMap<Scenario::MULTI_MAP, DIM>& tree, const Query& query) {
+size_t CountEntries(TestMap<Scenario::MULTI_MAP, DIM>& tree, const Query& query) {
     CounterMultiMap counter{query.center, query.radius, 0};
     tree.for_each(query.box, counter);
     return counter.n_;
 }
 
 template <dimension_t DIM, Scenario SCENARIO>
-int CountEntries(TestMap<Scenario::MULTI_MAP_STD, DIM>& tree, const Query& query) {
+size_t CountEntries(TestMap<Scenario::MULTI_MAP_STD, DIM>& tree, const Query& query) {
     CounterMultiMap counter{query.center, query.radius, 0};
     tree.for_each(query.box, counter);
     return counter.n_;
@@ -199,7 +199,7 @@ void IndexBenchmark<DIM, SCENARIO>::SetupWorld(benchmark::State& state) {
 
 template <dimension_t DIM, Scenario SCENARIO>
 void IndexBenchmark<DIM, SCENARIO>::QueryWorld(benchmark::State& state, const Query& query) {
-    int n = CountEntries<DIM, SCENARIO>(tree_, query);
+    size_t n = CountEntries<DIM, SCENARIO>(tree_, query);
 
     state.counters["query_rate"] += 1;
     state.counters["result_rate"] += n;
